add creator-feed and main-feed re-render after upload form submit
